requirements.txt
blinker==1.7.0
click==8.1.7
Flask==3.0.2
itsdangerous==2.1.2
Jinja2==3.1.3
Markdown==3.6
MarkupSafe==2.1.5
Werkzeug==3.0.1
torch==2.0.1
transformers
faiss-cpu
jupyter
ipykernel
Flask-Session
flask-sqlalchemy 
flask-bcrypt